combine all annotations utils
